refactor(agent): separate harness catalog from mcp API
