+++ packages/browseros-agent/crates/harness-integrations/src/mcp/mod.rs
@@ -1,0 +1,11 @@
+pub(crate) mod paths;
+pub(crate) mod types;
+
+pub use paths::{
+    detect_installed_agents, is_installed, resolve_agent_mcp_config_path, resolve_agent_surface,
+};
+pub use types::{
+    AgentInfo, AgentScope, AgentSurface, DisconnectInput, DisconnectSummary, LinkInput,
+    LinkSummary, ListLinksFilter, ListedLink, ManifestLinkEntry, ManifestServerEntry, McpServer,
+    McpServerSpec, RescanEntry, RescanReport, ServerManifest, UnlinkInput, UnlinkSummary,
+};
